rename Dub.co bot
